move providers to wire.go
